+++ src/main/java/ru/yandex/practicum/storage/like/LikeStorage.java
@@ -5,9 +5,9 @@ import ru.yandex.practicum.model.Film;
 import java.util.List;
 
 public interface LikeStorage {
-    void addLike(Long filmId, Long userId);
+    void addLike(Integer filmId, Integer userId);
 
-    void removeLike(Long filmId, Long userId);
+    void removeLike(Integer filmId, Integer userId);
 
     List<Film> getPopularFilms(int count);
 }
