+++ src/main/java/ru/yandex/practicum/storage/mpa/MpaStorage.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface MpaStorage {
     List<MpaRating> findAll();
 
-    Optional<MpaRating> findById(Long id);
+    Optional<MpaRating> findById(Integer id);
 }
